requirements.txt
scikit-learn
pandas
fastapi
uvicorn
python-multipart
python-dotenv
streamlit
mlflow